`extension/getLeanClient()` - make paperproof work with the latest `lean-vscode extension` version
