--- extension/src/services/Readme.md
@@ -1,1 +1,0 @@
-This is a folder for library-like methods - methods that could potentially be used in any vscode extension.
